+++ app/Http/Controllers/Admin/Advantage/AdvantageController.php
@@ -7,7 +7,7 @@ use App\Http\Resources\AdvantagesResource;
 use App\Http\Resources\AdvantegeOneResource;
 use App\Models\Advantage;
 use App\Http\Requests\AdvantageRequest;
-use App\Traits\ResponseTrait;
+use App\Http\Traits\ResponseTrait;
 
 
 class AdvantageController extends Controller
@@ -31,8 +31,11 @@ class AdvantageController extends Controller
     public function show($id)
     {
         $advantage = Advantage::find($id);
-        $response['advantage'] = new AdvantegeOneResource($advantage);
-        return self::okResponse($response);
+        if ($advantage) {
+            $response['advantage'] = new AdvantegeOneResource($advantage);
+            return self::okResponse($response);
+        }
+        return self::notFoundResponse();
     }
 
     /**
@@ -55,10 +58,13 @@ class AdvantageController extends Controller
     public function update(AdvantageRequest $request, $id)
     {
         $advantage = Advantage::find($id);
-        $data = $request->validated();
-        $advantage->update($data);
-        $response['advantage'] = new AdvantegeOneResource($advantage);
-        return self::okResponse($response);
+        if ($advantage) {
+            $data = $request->validated();
+            $advantage->update($data);
+            $response['advantage'] = new AdvantegeOneResource($advantage);
+            return self::okResponse($response);
+        }
+        return self::notFoundResponse();
     }
 
     /**
@@ -68,8 +74,11 @@ class AdvantageController extends Controller
     public function destroy($id)
     {
         $advantage = Advantage::find($id);
-        $advantage->delete();
-        return $this->index();
+        if ($advantage) {
+            $advantage->delete();
+            return $this->index();
+        }
+        return self::notFoundResponse();
     }
 }
